+++ src/core_modules/capture-core/actions/navigateToEnrollmentOverview/navigateToEnrollmentOverview.epics.js
@@ -5,14 +5,14 @@ import { switchMap } from 'rxjs/operators';
 import { EMPTY } from 'rxjs';
 import { config } from 'd2';
 import { actionTypes as NavigateToEnrollmentOverviewActionTypes } from './navigateToEnrollmentOverview.actions';
-import { buildUrlQueryString } from '../../utils/routing';
+import { buildUrlQueryString, deriveURLParamsFromLocation } from '../../utils/routing';
 import { scopeHierarchyTypes } from './navigateToEnrollmentOverview.constants';
 
 // TODO This will be removed when the link between capture and tracker capture is not relevant
-const redirectToTracker = ({ teiId, orgUnitId, dependencies, store }) => {
+const redirectToTracker = ({ teiId, orgUnitId, dependencies }) => {
     const { baseUrl } = config;
     const { search, pathname } = dependencies.history.location;
-    const { programId: queryProgramId, trackedEntityTypeId: queryTrackedEntityTypeId } = store.value.router.location.query;
+    const { programId: queryProgramId, trackedEntityTypeId: queryTrackedEntityTypeId } = deriveURLParamsFromLocation();
 
     const instanceBaseUrl = baseUrl.split('/api')[0];
     const scopeHierarchy = queryProgramId ? scopeHierarchyTypes.PROGRAM : scopeHierarchyTypes.TRACKED_ENTITY_TYPE;
@@ -35,27 +35,26 @@ const redirectToEnrollmentDashboard = ({ dependencies, teiId, programId, orgUnit
     );
 };
 
+const shouldUseNewDashboard = (userDataStore, dataStore, programId) =>
+    userDataStore?.[programId] || (userDataStore?.[programId] !== false && dataStore?.[programId]);
+
 export const navigateToEnrollmentOverviewEpic = (action$: InputObservable, store: ReduxStore, dependencies: any) =>
     action$.pipe(
         ofType(NavigateToEnrollmentOverviewActionTypes.NAVIGATE_TO_ENROLLMENT_OVERVIEW),
         switchMap((action) => {
             const { teiId, programId, orgUnitId } = action.payload;
             const enrollmentId = programId && (action.payload?.enrollmentId || 'AUTO');
-            const { dataStore, userDataStore } = store.value.useOldDashboard;
+            const { dataStore, userDataStore } = store.value.useNewDashboard;
 
             if (dataStore || userDataStore) {
-                if (userDataStore?.[programId]) {
-                    redirectToTracker({ dependencies, store, teiId, orgUnitId });
-                    return EMPTY;
-                }
-
-                if (userDataStore?.[programId] !== false && dataStore?.[programId]) {
-                    redirectToTracker({ dependencies, store, teiId, orgUnitId });
+                const shouldRedirectToEnrollmentDashboard = shouldUseNewDashboard(userDataStore, dataStore, programId);
+                if (shouldRedirectToEnrollmentDashboard) {
+                    redirectToEnrollmentDashboard({ dependencies, teiId, programId, orgUnitId, enrollmentId });
                     return EMPTY;
                 }
             }
 
-            redirectToEnrollmentDashboard({ dependencies, teiId, programId, orgUnitId, enrollmentId });
+            redirectToTracker({ dependencies, store, teiId, orgUnitId });
             return EMPTY;
         }),
     );