+++ src/core_modules/capture-core/components/DataStore/DataStore.actions.js
@@ -1,7 +1,7 @@
 // @flow
 import { actionCreator } from '../../actions/actions.utils';
-import { type UseOldDashboard, actionTypes as DataStoreActionTypes } from './DataStore.types';
+import { type UseNewDashboard, actionTypes as DataStoreActionTypes } from './DataStore.types';
 
 
-export const fetchDataStore = ({ dataStore, userDataStore }: UseOldDashboard) =>
+export const fetchDataStore = ({ dataStore, userDataStore }: UseNewDashboard) =>
     actionCreator(DataStoreActionTypes.FETCH_DATA_STORE)({ dataStore, userDataStore });