+++ src/core_modules/capture-core/reducers/descriptions/dataStore.reducerDescription.js
@@ -15,7 +15,7 @@ export const dataStoreDesc = createReducerDescription({
 
         return newState;
     },
-}, 'useOldDashboard', {
+}, 'useNewDashboard', {
     dataStore: undefined,
     userDataStore: undefined,
 });